Renamed Measurement.py to MeasurementSofastFringe.py.
