+++ contrib/test_data_generation/sofast/generate_test_data_undefined.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from opencsp.app.sofast.lib.ImageCalibrationScaling import ImageCalibrationScaling
-from opencsp.app.sofast.lib.Measurement import Measurement
+from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.app.sofast.lib.Sofast import Sofast
 from opencsp.common.lib.deflectometry.Display import Display
 from opencsp.common.lib.camera.Camera import Camera
