+++ opencsp/app/sofast/calibration/test/test_CalibrationScreenShape.py
@@ -12,7 +12,7 @@ from opencsp.app.sofast.calibration.lib.CalibrationScreenShape import (
     CalibrationScreenShape,
     DataInput,
 )
-from opencsp.app.sofast.lib.Measurement import Measurement
+from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.common.lib.camera.Camera import Camera
 from opencsp.common.lib.deflectometry.ImageProjection import ImageProjection
 from opencsp.common.lib.geometry.Vxyz import Vxyz
