+++ opencsp/app/sofast/lib/Sofast.py
@@ -6,7 +6,7 @@ import warnings
 
 import numpy as np
 
-from opencsp.app.sofast.lib.Measurement import Measurement
+from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.app.sofast.lib.SofastParams import SofastParams
 from opencsp.common.lib.camera.Camera import Camera
 from opencsp.common.lib.csp.Facet import Facet
