+++ opencsp/app/sofast/test/test_ImageCalibrationGlobal.py
@@ -5,7 +5,7 @@ import datetime as dt
 import numpy as np
 
 from opencsp.app.sofast.lib.ImageCalibrationGlobal import ImageCalibrationGlobal
-from opencsp.app.sofast.lib.Measurement import Measurement
+from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.common.lib.geometry.Vxyz import Vxyz
