+++ opencsp/app/sofast/test/test_integration_multi_facet.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from opencsp.app.sofast.lib.ImageCalibrationScaling import ImageCalibrationScaling
 from opencsp.app.sofast.lib.Sofast import Sofast
-from opencsp.app.sofast.lib.Measurement import Measurement
+from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.common.lib.camera.Camera import Camera
 from opencsp.common.lib.deflectometry.Display import Display
 from opencsp.common.lib.deflectometry.EnsembleData import EnsembleData
