+++ opencsp/app/sofast/test/test_integration_single_facet.py
@@ -7,7 +7,7 @@ import unittest
 import numpy as np
 
 from opencsp.app.sofast.lib.ImageCalibrationScaling import ImageCalibrationScaling
-from opencsp.app.sofast.lib.Measurement import Measurement
+from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.app.sofast.lib.Sofast import Sofast
 from opencsp.common.lib.camera.Camera import Camera
 from opencsp.common.lib.deflectometry.Display import Display
