+++ opencsp/app/sofast/test/test_integration_undefined.py
@@ -5,7 +5,7 @@ import os
 import numpy as np
 
 from opencsp.app.sofast.lib.ImageCalibrationScaling import ImageCalibrationScaling
-from opencsp.app.sofast.lib.Measurement import Measurement
+from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.app.sofast.lib.Sofast import Sofast
 from opencsp.common.lib.camera.Camera import Camera
 from opencsp.common.lib.deflectometry.Display import Display
